requirements.txt
aartfaac-arthur>=0.1,<0.3
django>=1.10,<1.11
gunicorn>=19.6,<19.7

